requirements.txt
asgiref==3.5.2
attrs==22.1.0
black==22.10.0
certifi==2022.9.24
charset-normalizer==2.1.1
click==8.1.3
colorama==0.4.6
Django==3.2.16
django-filter==21.1
djangorestframework==3.14.0
exceptiongroup==1.0.4
factory-boy==3.2.1
Faker==15.3.4
flake8==4.0.1
flake8-bugbear==22.10.27
flake8-import-order==0.18.1
idna==3.4
iniconfig==1.1.1
mccabe==0.6.1
mypy-extensions==0.4.3
mysqlclient==2.1.1
packaging==21.3
pathspec==0.10.2
platformdirs==2.5.4
pluggy==1.0.0
pycodestyle==2.8.0
pyflakes==2.4.0
pyparsing==3.0.9
pytest==7.2.0
pytest-django==4.5.2
pytest-flake8==1.1.1
python-dateutil==2.8.2
python-dotenv==0.20.0
pytz==2022.6
requests==2.28.1
six==1.16.0
sqlparse==0.4.3
tomli==2.0.1
typing_extensions==4.4.0
tzdata==2022.7
urllib3==1.26.13
uuid==1.30
